ozondrf/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.5
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-timezone-field==5.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
flower==1.2.0
gunicorn==20.1.0
humanize==4.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
loguru==0.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
py3dbp==1.1.2
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
redis==4.4.2
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
supervisor==4.2.5
tornado==6.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
